ユーザーアイコンのMineを正しく設定されるように修正
