+++ app/src/main/java/intern/line/me/kyotoaclient/presenter/room/PostRoomIcon.kt
@@ -13,6 +13,7 @@ import retrofit2.Response
 import ru.gildor.coroutines.retrofit.awaitResponse
 import java.io.File
 import android.graphics.BitmapFactory
+import intern.line.me.kyotoaclient.lib.util.IconFiles
 
 class PostRoomIcon: API() {
 
@@ -20,7 +21,7 @@ class PostRoomIcon: API() {
 
 
 	suspend fun postAsyncRoomIcon(room_id: Long, token: String, file: File): Response<Boolean> = withContext(CommonPool) {
-		val mime = getMimeTypeOfFile(file.absolutePath)
+		val mime = IconFiles().getMimeTypeOfFile(file.absolutePath)
 		val requestBody: RequestBody = RequestBody
 				.create(MediaType.parse(mime), file)
 
@@ -47,12 +48,5 @@ class PostRoomIcon: API() {
 			}
 		}
 		return false
-	}
-
-	fun getMimeTypeOfFile(pathName: String): String {
-		val opt = BitmapFactory.Options()
-		opt.inJustDecodeBounds = true
-		BitmapFactory.decodeFile(pathName, opt)
-		return opt.outMimeType
 	}
 }
